rename 'voting' --> 'votes'
