+++ src/redux/features/votes/votesSelectors.ts
