+++ src/types/Votes.ts
